requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
supabase==2.0.3
pytest==7.4.3
httpx==0.25.2
psycopg2-binary==2.9.9  # For database operations
requests==2.31.0 